python/многопоточность/tornado/tornado_and_django/requirements.txt
Django==1.9.5
django-redis-sessions==0.5.0
